refactor(gates): data-drive TOOL_CATEGORIES + lift HookContext/tool_categories to lib/
Two related cleanups in the gate plumbing:

1. **TOOL_CATEGORIES drift fix.** The 258-entry hand-maintained set was
   scanned O(n) per PreToolUse (docstring falsely claimed O(1)), carried
   an empty `'meta': set()`, and diverged from `_PKB_OPERATIONS` (16 ops
   only in the fallback, 1 only in the static set). The new module makes
   `_PKB_OPERATIONS` the single source: every prefix variant is generated
   mechanically at module load, an O(1) reverse `{tool->category}` index
   is built once, and `meta` is gone.

2. **Layer-violation fix.** `lib/gates/{engine,custom_conditions,custom_actions}.py`
   imported upward from `hooks/` (`hooks.schemas.HookContext`,
   `hooks.gate_config.get_tool_category/is_never_block`), with deferred
   in-method imports masking the issue. `HookContext` moves to
   `lib/hook_context.py`; `TOOL_CATEGORIES`, `get_tool_category`,
   `is_never_block`, and the COMPLIANCE/SPAWN registries move to
   `lib/tool_categories.py`. `hooks/schemas.py` and `hooks/gate_config.py`
   re-export for backward compatibility — gate-mode env-var resolution
   and slash-command/enforcer-channel hygiene stay in `hooks/gate_config.py`.

AC3 (factor copy-pasted block+warn `GatePolicy` pairs) is marked
optional in the brief and deferred to a follow-up.

Tests: 1452 hooks+lib tests pass. The two `test_handover_gate_latch`
failures (`test_interactive_*_closes_gate`) reproduce on `main` and are
unrelated to this refactor.

Closes epic-2bafa773
