+++ aops-core/hooks/schemas.py
@@ -1,60 +1,21 @@
 from typing import Any, Literal, TypeAlias
 
-from pydantic import BaseModel, ConfigDict, Field
+# HookContext lives in lib/ so gate-engine code in lib/gates/* can consume it
+# without importing upward from hooks/. Re-exported here for backward
+# compatibility with downstream callers.
+from lib.hook_context import HookContext
+from pydantic import BaseModel, Field
 
-# --- Input Schemas (Context) ---
-
-
-class HookContext(BaseModel):
-    """
-    Normalized input context for all hooks.
-
-    Precomputed values (session_short_hash, is_subagent) are computed once
-    during normalize_input() to avoid redundant calculations across gates.
-    """
-
-    model_config = ConfigDict()
-
-    # Core Identity
-    session_id: str = Field(..., description="The unique session identifier.")
-    trace_id: str | None = Field(
-        None, description="The unique ID for the specific hook invocation (tracing)."
-    )
-    hook_event: str = Field(
-        ..., description="The normalized event name (e.g., SessionStart, PreToolUse)."
-    )
-    agent_id: str | None = None
-    slug: str | None = None
-    client_type: str | None = None  # "claude" or "gemini" (set from --client flag)
-
-    # Metadata (aops-d9ba7159)
-    machine: str | None = None
-    provider: str | None = None
-    crew: str | None = None
-    repo: str | None = None
-    task_id: str | None = None
-
-    # Precomputed values (computed once in router.normalize_input())
-    session_short_hash: str = Field(
-        default="", description="8-char hash of session_id (computed once at normalization)."
-    )
-    is_subagent: bool = Field(
-        default=False,
-        description="Whether this is a subagent session (computed once at normalization).",
-    )
-
-    # Event Data
-    tool_name: str | None = None
-    tool_input: dict[str, Any] | list[Any] = Field(default_factory=dict)
-    tool_output: dict[str, Any] | list[Any] = Field(default_factory=dict)
-
-    transcript_path: str | None = None
-    cwd: str | None = None
-
-    subagent_type: str | None = None
-
-    # Raw Input (for fallback/passthrough)
-    raw_input: dict[str, Any] = Field(default_factory=dict)
+__all__ = [
+    "HookContext",
+    "ClaudeHookSpecificOutput",
+    "ClaudeStopHookOutput",
+    "ClaudeGeneralHookOutput",
+    "ClaudeHookOutput",
+    "GeminiHookSpecificOutput",
+    "GeminiHookOutput",
+    "CanonicalHookOutput",
+]
 
 
 # --- Claude Code Hook Schemas ---
